feat: day 6
This commit also changes the framework functions in `lib.rs` to use `u64` instead of `u32`.

+++ src/lib.rs
@@ -7,7 +7,7 @@ fn read_input_file(day: u8, initials: &str) -> String {
         .expect("You typo'd a filename (or sutin).")
 }
 
-pub fn solve_day(day: u8, part1: fn(&str) -> u32, part2: fn(&str) -> u32) {
+pub fn solve_day(day: u8, part1: fn(&str) -> u64, part2: fn(&str) -> u64) {
     println!("===== Begin Part 1 =====");
     INITIALS.iter().for_each(|&initials| {
         let input = read_input_file(day, initials);
